simplify model and remove batch norm for now
I realized that my batch normalization package is incorrect, so it doesn't help the network at all. I will try adding it back once it is correct. In the meantime, I experimented with getting the network to run in single in-memory mini-batches, because that will be necessary for correct batch normalization.